Transcode unsupported audio formats to Ogg
Transcoding is handled on the backend with avconv from libav.

The client requires songs via a new endpoint:
    GET /song/<id>/<wanted_formats>
where id is the integer id of the song in the database,
and wanted_formats is a comma-separated string of file extensions,
e.g.:
    GET /song/491/ogg,mp3,m4a

If wanted_formats includes the file's original format, it's sent
without transcoding.

Otherwise, if wanted_formats include ogg, it's transcoded to ogg.
No caching is done yet.