Fix bug: re-adding album corrupted playlist data
Steps to reproduce:
1. Add a track (that is on an album) to the playlist
2. Add that track's album
3. Try to delete the track

Expected result:
The track is deleted.

Actual result before this commit:
The track still shows in the list. Further attempts to delete it print
an error on the console: "tried to remove song not in playlist!"

This was happening because, when an album is added, the collection's
`.set()` helpfully "merged" new values for the track, which then didn't
match the HTML id stored in the DOM. The fix is to set `merge: false`
for the call to `set()`.