Include only minimal necessary code on login page
Previously we were sending the entire app JS bundle with the login page,
which didn't make sense, and led to a bug where the default in-app
background image would sometimes be applied to the login page.

I rewrote Flask-BrowserID's JS include to remove the jQuery dependency,
so now we only refer to Persona's include and ~50 lines of rewritten
code to submit assertions to the backend API and reload the page.