Give Backbone an explicit reference to jQuery
This is a hack, but fixes an error that mysteriously appeared in Pots
without any of my code changing. Possibly I upgraded my globally
installed version of Browserify.

Anyway, the root problem is that Backbone is not CommonJS-aware, and
attempts to save a global reference to jQuery, which will only exist if
jQuery has been defined (not imported!) before Backbone.