Flux rewrite with new design
Fix #39 (new design)

Also: Fix #41 (play/pause bugs)

A notable backend component is this changes the search endpoint to
return flat results, with all tracks/albums referenced included. (But
not siblings. If track T from album A is included, then A's metadata is
included, but the metadata for A's other tracks is *not* included. On
the other hand, if album A itself is included in the search results, all
of its tracks' metadata are included.) This may or may not have broken
the /album/:id and /track/:id endpoints, but it also obviates the need
to use them for now. :)